Uradjen i drugi deo zadatka: Upravljanje izuzecima.

+++ BookstoreApplication/Services/Implementations/AuthorService.cs
@@ -1,4 +1,5 @@
-﻿using BookstoreApplication.Models;
+﻿using BookstoreApplication.Exceptions;
+using BookstoreApplication.Models;
 using BookstoreApplication.Repositories.Interfaces;
 using BookstoreApplication.Services.Interfaces;
 
@@ -25,8 +26,8 @@ namespace BookstoreApplication.Services.Implementations
 
         public async Task UpdateAsync(int id, Author dto, CancellationToken ct = default)
         {
-            if (id != dto.Id) throw new ArgumentException("ID mismatch.");
-            if (!await _authors.ExistsAsync(id)) throw new KeyNotFoundException();
+            if (id != dto.Id) throw new BadRequestException("ID mismatch.");
+            if (!await _authors.ExistsAsync(id)) throw new NotFoundException(id);
 
             await _authors.UpdateAsync(dto);
             await _authors.SaveChangesAsync();
@@ -34,6 +35,7 @@ namespace BookstoreApplication.Services.Implementations
 
         public async Task DeleteAsync(int id, CancellationToken ct = default)
         {
+            if (!await _authors.ExistsAsync(id)) throw new NotFoundException(id);
             await _authors.DeleteAsync(id);
             await _authors.SaveChangesAsync();
         }
